continue the 'accessMiddleware' refactor
