+++ src/routes/v1/financial/withdrawals.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 import { withdrawalAdminController, workerEarningsController } from '../../../state.js';
 import { authenticateAccess, isActive } from '../../../middlewares/authMiddleware.js';
 import { validateBody, validateParams } from '../../../middlewares/validateRequest.js';
-import { authorizeAdmin } from '../../../middlewares/adminMiddleware.js';
+import { authorizeAdmin } from '../../../middlewares/accessMiddleware.js';
 import { z } from 'zod';
